+++ servers/src/common/stats.rs
@@ -26,7 +26,7 @@ use crate::core::ser::ProtocolVersion;
 use chrono::prelude::*;
 
 use crate::chain;
-use crate::common::types::SyncStatus;
+use crate::chain::SyncStatus;
 use crate::p2p;
 
 /// Server state info collection struct, to be passed around into internals
